Prevent delete of in use environment

+++ pkg/server/handler/environment.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"errors"
 	"strconv"
 
 	"github.com/productiveops/dokemon/pkg/server/model"
@@ -80,6 +81,15 @@ func (h *Handler) DeleteEnvironmentById(c echo.Context) error {
 
 	if !exists {
 		return resourceNotFound(c, "Environment")
+	}
+
+	inUse, err := h.environmentStore.IsInUse(uint(id))
+	if err != nil {
+		panic(err)
+	}
+
+	if inUse {
+		return unprocessableEntity(c, errors.New("Environment is in use and cannot be deleted"))
 	}
 
 	if err := h.environmentStore.DeleteById(uint(id)); err != nil {
